requirements.txt
# Core ML & NLP
faiss-cpu==1.7.4
sentence-transformers==2.5.1
numpy==1.26.4

# Backend - FastAPI
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1

# LLM - Google Gemini API
google-generativeai==0.5.2

# Frontend - Streamlit
streamlit==1.44.1

# Web Requests & Utils
requests==2.31.0

# Optional - for eval and pretty display
pandas==2.2.2
